Use IP address as key, user-friendly comments, param check
Previously we use truncated domain names as key, which may lead to
    192.168.1.2 x.aaa.com
    192.168.2.2 y.aaa.com
being converted to
    address=/aaa.com/192.168.2.2

This is obviously wrong. This commit will change it to
    address=/x.aaa.com/192.168.1.2
    address=/y.aaa.com/192.168.2.2

Note that
    192.168.3.2 cdn1.aaa.com
    192.168.3.2 cdn2.aaa.com
will be converted to
    address=/aaa.com/192.168.3.2
If www.aaa.com uses different IP from its CDN,
this site will be unusable. Don't forget to check output carefully!

Other improvements:
 * Shows built-in README when parameters are missing
 * Shows error on IO errors: file not found, no permission etc
 * Make logic, comments, names clearer